OpenShift tests packages refactoring
