+++ tests/org.jboss.tools.openshift.ui.bot.test/src/org/jboss/tools/openshift/ui/bot/test/TestCasesTestSuite.java
@@ -1,14 +1,6 @@
 package org.jboss.tools.openshift.ui.bot.test;
 
 import org.jboss.reddeer.junit.runner.RedDeerSuite;
-import org.jboss.tools.openshift.application.advanced.ID904DeployApplicationWARArchiveTest;
-import org.jboss.tools.openshift.application.advanced.ID905ManageSnapshotsTest;
-import org.jboss.tools.openshift.ui.bot.test.application.basic.ID306PreselectLastUsedConnectionTest;
-import org.jboss.tools.openshift.ui.bot.test.application.basic.ID307WizardDataProcessingTest;
-import org.jboss.tools.openshift.ui.bot.test.application.cartridge.ID601EmbedCartridgeTest;
-import org.jboss.tools.openshift.ui.bot.test.application.create.ID406CreateApplicationOnDownloadableCartridgeTest;
-import org.jboss.tools.openshift.ui.bot.test.application.create.ID417CreateApplicationWithEnvironmentVariablesTest;
-import org.jboss.tools.openshift.ui.bot.test.application.importing.ID503ImportApplicationViaAdapterTest;
 import org.jboss.tools.openshift.ui.bot.test.common.ID001RemoteRequestTimeoutTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.ID103xCreateNewConnectionEnterpriseTest;
 import org.jboss.tools.openshift.ui.bot.test.connection.ID108xHandleMoreServersTest;
@@ -39,9 +31,6 @@ import org.junit.runners.Suite.SuiteClasses;
 
 	ID201NewDomainTest.class,
 	ID204CreateMoreDomainsTest.class,
-	
-	ID503ImportApplicationViaAdapterTest.class,
-	ID905ManageSnapshotsTest.class,
 	
 	CleanUp.class})
 public class TestCasesTestSuite {
