+++ tests/org.jboss.tools.openshift.ui.bot.test/src/org/jboss/tools/openshift/ui/bot/test/application/adapter/ID803ServerAdapterHandlingTest.java
@@ -18,8 +18,8 @@ import org.jboss.reddeer.swt.impl.tree.DefaultTree;
 import org.jboss.reddeer.swt.wait.TimePeriod;
 import org.jboss.reddeer.swt.wait.WaitUntil;
 import org.jboss.reddeer.swt.wait.WaitWhile;
-import org.jboss.tools.openshift.application.advanced.ID903ApplicationMarkersTest;
-import org.jboss.tools.openshift.application.advanced.ID905ManageSnapshotsTest;
+import org.jboss.tools.openshift.ui.bot.test.application.advanced.ID903ApplicationMarkersTest;
+import org.jboss.tools.openshift.ui.bot.test.application.advanced.ID905ManageSnapshotsTest;
 import org.jboss.tools.openshift.ui.bot.test.application.cartridge.ID601EmbedCartridgeTest;
 import org.jboss.tools.openshift.ui.bot.test.application.create.ID408ApplicationPropertiesTest;
 import org.jboss.tools.openshift.ui.bot.test.application.handle.ID705TailFilesTest;
