+++ tests/org.jboss.tools.openshift.ui.bot.test/src/org/jboss/tools/openshift/ui/bot/test/application/advanced/ID903ApplicationMarkersTest.java
@@ -1,4 +1,4 @@
-package org.jboss.tools.openshift.application.advanced;
+package org.jboss.tools.openshift.ui.bot.test.application.advanced;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
